requirements.txt
psycopg2-binary
apache-airflow==2.9.0
pandas